requirements.txt
numpy==1.20.0
opencv-python==4.5.1.48
Pillow==8.1.0
tensorflow-gpu==1.15.0
Keras==2.2.4
